WIP: split the locking suite and share its fixture
